--- app/components/ProductCard/AddToCart.tsx
@@ -1,14 +1,0 @@
-"use client"
-
-const AddToCart = () => {
-  return (
-    <button
-      className="btn btn-primary"
-      onClick={() => console.log("Add to cart clicked")}
-    >
-      Add to cart
-    </button>
-  )
-}
-
-export default AddToCart
