--- app/components/ProductCard/ProductCard.tsx
@@ -1,11 +1,0 @@
-import AddToCart from "./AddToCart"
-
-const ProductCard = () => {
-  return (
-    <div>
-      <AddToCart />
-    </div>
-  )
-}
-
-export default ProductCard
