+++ app/components/atoms/index.ts
@@ -1,0 +1,4 @@
+import Button from "./Button"
+import Loading from "./loading"
+
+export { Button, Loading }
