+++ app/components/index.ts
@@ -1,0 +1,2 @@
+export * from "./atoms"
+export * from "./molecules"
