+++ app/components/molecules/NavBar/index.tsx
@@ -1,6 +1,6 @@
 import Link from "next/link"
 
-const NavBAr = () => {
+const NavBar = () => {
   return (
     <div className="flex bg-slate-200 p-5">
       <Link href="/" className="mr-5">
@@ -19,4 +19,4 @@ const NavBAr = () => {
   )
 }
 
-export default NavBAr
+export default NavBar
