+++ app/components/molecules/ProductCard/AddToCartButton/index.tsx
@@ -1,4 +1,4 @@
-import Button from "../../Button"
+import { Button } from "@/app/components"
 
 const AddToCartButton = () => {
   return <Button onClick={() => alert("Add to cart")}>Add to cart</Button>
