+++ app/components/molecules/ProductCard/index.tsx
@@ -6,6 +6,7 @@ import styles from "./styles.module.css"
 const ProductCard = () => {
   return (
     <div className={styles.card}>
+      <h2>Product</h2>
       <AddToCartButton />
     </div>
   )
