+++ app/components/molecules/index.ts
@@ -1,0 +1,5 @@
+import NavBar from "./NavBar"
+import ProductCard from "./ProductCard"
+import UserTable from "./UserTable"
+
+export { NavBar, ProductCard, UserTable }
