+++ app/layout.tsx
@@ -3,7 +3,7 @@ import "./globals.css"
 import type { Metadata } from "next"
 import { Inter } from "next/font/google"
 
-import NavBAr from "./NavBar"
+import { NavBar } from "@/app/components"
 
 const inter = Inter({ subsets: ["latin"] })
 
@@ -20,7 +20,7 @@ export default function RootLayout({
   return (
     <html lang="en" data-theme="winter">
       <body className={inter.className}>
-        <NavBAr />
+        <NavBar />
         <main className="p-5"> {children}</main>
       </body>
     </html>
